Rework the model to be more efficient
The new approach works by creating two cell shapes--one with a small hole
and one without--and then just repeating those shapes enough times to get
a tile.

OpenSCAD seems to like this approach a lot better.  In a lot of tests, it
cut rendering time almost in half.

This should also make it easier to render non-rectangular tiles in the
future.