Flare the ends of the peg hole, just like the STEP models
Note that older Multiboard tile STLs have straight-sided peg holes, like
this model used to.  At some point, the design was revised, the STEP files
were updated, and all new models with peg holes got the flares, but the
old STLs were left as they were.

I also renamed a bunch of variables to make it easier for me to track what
everything was for.  This is mixing two different changes, but the new
names helped me with my work on the peg holes, and it would have been too
much work to detangle the changes.